Mueve de arichivo el componente ImageCard
